backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
jinja2>=3.1.0
playwright>=1.40.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
requests>=2.31.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
